Move from webpack to vite, general clean up
Yes this commit is extremely large, the gist of it is just getting rid
of the extremely painful webpack configuration stuff in favor of vite
(which also makes builds wayyy fater) and getting it to actually compile.
Besides for that, this commit is mostly updating old packages and
converting from class components to function components(as well as
some other more modern practices).

The main goal was getting off of webpack and making everything compile
so there are still more changes I plan to make. Namely using react-query
to manage server data and either updating or switching off react-router.